Tested Heathrow
Added tests for the image input. Doesn't like it however as its pulling
the background image of the panel instead of the drawn on image.
Created toString methods in Waypoint and Position so that the object
Waypoint can be comparable. Still failed however - not sure why this is
Updated Unresolved failures